Finish rename of QCJSON -> QCSchema
