+++ cclib/io/__init__.py
@@ -17,8 +17,8 @@ from cclib.io.cjsonreader import CJSON as CJSONReader
 from cclib.io.cjsonwriter import CJSON as CJSONWriter
 from cclib.io.cmlwriter import CML
 from cclib.io.moldenwriter import MOLDEN
-from cclib.io.qcjsonreader import QCSchemaReader
-from cclib.io.qcjsonwriter import QCSchemaWriter
+from cclib.io.qcschemareader import QCSchemaReader
+from cclib.io.qcschemawriter import QCSchemaWriter
 from cclib.io.wfxwriter import WFXWriter
 from cclib.io.xyzreader import XYZ as XYZReader
 from cclib.io.xyzwriter import XYZ as XYZWriter
