+++ cclib/io/qcschemareader.py
